WinNUKE is a modified version of LOIC
WinNUKE is my attempt to take the greatness from LOIC and remove the
skiddie graphics and prompts. I give all credit to the author, as we all
know there are not many great Network stress testing tools on Windows
platform. I aim to make this program a sleek go-to solution for windows
users and network admin.